+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning].
 
 - RandomSkunk.Results.AspNetCore:
     - Add `ToFileActionResult` extension methods.
+    - Add `ToActionResult` overloads that have an `onSuccess` function parameter.
 
 ### Fixed
 - RandomSkunk.Results: